pyproject.toml
[tool.poetry]
name = "url_shortner_backend"
version = "0.1.0"
description = ""
authors = ["lokaimoma <kelvinclark2k77@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.1.0"
python-dotenv = "^0.20.0"
django-cors-headers = "^3.11.0"
PyMySQL = "^1.0.2"
shortuuid = "^1.0.8"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.1"
Faker = "^13.3.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
